refactor(oidc): Get rid of OidcBackend
Now that we don't use it for tests, we don't need it anymore.
