+++ crates/matrix-sdk/src/authentication/oidc/mod.rs
@@ -146,22 +146,31 @@
 //! [`AuthenticateError::InsufficientScope`]: ruma::api::client::error::AuthenticateError
 //! [`examples/oidc_cli`]: https://github.com/matrix-org/matrix-rust-sdk/tree/main/examples/oidc_cli
 
-use std::{collections::HashMap, fmt, sync::Arc};
+use std::{collections::HashMap, fmt, future::Future, pin::Pin, sync::Arc};
 
 use as_variant::as_variant;
+use chrono::Utc;
 use eyeball::SharedObservable;
 use futures_core::Stream;
 pub use mas_oidc_client::{error, requests, types};
 use mas_oidc_client::{
+    http_service::HttpService,
     requests::{
         account_management::{build_account_management_url, AccountManagementActionFull},
-        authorization_code::AuthorizationValidationData,
+        authorization_code::{access_token_with_authorization_code, AuthorizationValidationData},
+        discovery::{discover, insecure_discover},
+        refresh_token::refresh_access_token,
+        registration::register_client,
+        revocation::revoke_token,
     },
     types::{
         client_credentials::ClientCredentials,
         errors::{ClientError, ClientErrorCode::AccessDenied},
         iana::oauth::OAuthTokenTypeHint,
-        oidc::{AccountManagementAction, VerifiedProviderMetadata},
+        oidc::{
+            AccountManagementAction, ProviderMetadata, ProviderMetadataVerificationError,
+            VerifiedProviderMetadata,
+        },
         registration::{ClientRegistrationResponse, VerifiedClientMetadata},
         requests::Prompt,
         scope::{MatrixApiScopeToken, ScopeToken},
@@ -170,7 +179,9 @@ use mas_oidc_client::{
 #[cfg(feature = "e2e-encryption")]
 use matrix_sdk_base::crypto::types::qr_login::QrCodeData;
 use matrix_sdk_base::{once_cell::sync::OnceCell, SessionMeta};
+use oauth2::{AsyncHttpClient, HttpClientError, HttpRequest, HttpResponse};
 use rand::{rngs::StdRng, Rng, SeedableRng};
+use ruma::api::client::discovery::{get_authentication_issuer, get_authorization_server_metadata};
 use serde::{Deserialize, Serialize};
 use sha2::Digest as _;
 use thiserror::Error;
@@ -179,7 +190,6 @@ use tracing::{debug, error, info, instrument, trace, warn};
 use url::Url;
 
 mod auth_code_builder;
-mod backend;
 mod cross_process;
 #[cfg(all(feature = "e2e-encryption", not(target_arch = "wasm32")))]
 pub mod qrcode;
@@ -192,13 +202,14 @@ pub use self::{
     cross_process::CrossProcessRefreshLockError,
 };
 use self::{
-    backend::{server::OidcServer, OidcBackend},
     cross_process::{CrossProcessRefreshLockGuard, CrossProcessRefreshManager},
     qrcode::LoginWithQrCode,
     registrations::{ClientId, OidcRegistrations},
 };
 use super::AuthData;
-use crate::{client::SessionChange, Client, HttpError, RefreshTokenError, Result};
+use crate::{
+    client::SessionChange, http_client::HttpClient, Client, HttpError, RefreshTokenError, Result,
+};
 
 pub(crate) struct OidcCtx {
     /// Lock and state when multiple processes may refresh an OIDC session.
@@ -251,18 +262,23 @@ impl fmt::Debug for OidcAuthData {
 pub struct Oidc {
     /// The underlying Matrix API client.
     client: Client,
-
-    /// The implementation of the OIDC backend.
-    backend: Arc<dyn OidcBackend>,
 }
 
 impl Oidc {
     pub(crate) fn new(client: Client) -> Self {
-        Self { client: client.clone(), backend: Arc::new(OidcServer::new(client)) }
+        Self { client }
     }
 
     fn ctx(&self) -> &OidcCtx {
         &self.client.inner.auth_ctx.oidc
+    }
+
+    fn http_client(&self) -> &HttpClient {
+        &self.client.inner.http_client
+    }
+
+    fn http_service(&self) -> HttpService {
+        HttpService::new(self.http_client().clone())
     }
 
     /// Enable a cross-process store lock on the state store, to coordinate
@@ -630,7 +646,55 @@ impl Oidc {
     /// Returns an error if a problem occurred when fetching or validating the
     /// metadata.
     pub async fn provider_metadata(&self) -> Result<VerifiedProviderMetadata, OauthDiscoveryError> {
-        self.backend.discover(self.ctx().insecure_discover).await
+        match self.client.send(get_authorization_server_metadata::msc2965::Request::new()).await {
+            Ok(response) => {
+                let metadata = response.metadata.deserialize_as::<ProviderMetadata>()?;
+
+                let result = if self.ctx().insecure_discover {
+                    metadata.insecure_verify_metadata()
+                } else {
+                    // The mas-oidc-client method needs to compare the issuer for validation. It's a
+                    // bit unnecessary because we take it from the metadata, oh well.
+                    let issuer =
+                        metadata.issuer.clone().ok_or(error::DiscoveryError::Validation(
+                            ProviderMetadataVerificationError::MissingIssuer,
+                        ))?;
+                    metadata.validate(&issuer)
+                };
+
+                return Ok(result.map_err(error::DiscoveryError::Validation)?);
+            }
+            Err(error)
+                if error
+                    .as_client_api_error()
+                    .is_some_and(|err| err.status_code == http::StatusCode::NOT_FOUND) =>
+            {
+                // Fallback to OIDC discovery.
+            }
+            Err(error) => return Err(error.into()),
+        };
+
+        // TODO: remove this fallback behavior when the metadata endpoint has wider
+        // support.
+        #[allow(deprecated)]
+        let issuer =
+            match self.client.send(get_authentication_issuer::msc2965::Request::new()).await {
+                Ok(response) => response.issuer,
+                Err(error)
+                    if error
+                        .as_client_api_error()
+                        .is_some_and(|err| err.status_code == http::StatusCode::NOT_FOUND) =>
+                {
+                    return Err(OauthDiscoveryError::NotSupported);
+                }
+                Err(error) => return Err(error.into()),
+            };
+
+        if self.ctx().insecure_discover {
+            insecure_discover(&self.http_service(), &issuer).await.map_err(Into::into)
+        } else {
+            discover(&self.http_service(), &issuer).await.map_err(Into::into)
+        }
     }
 
     /// The OpenID Connect unique identifier of this client obtained after
@@ -832,10 +896,13 @@ impl Oidc {
             .as_ref()
             .ok_or(OidcError::NoRegistrationSupport)?;
 
-        let registration_response = self
-            .backend
-            .register_client(registration_endpoint, client_metadata, software_statement)
-            .await?;
+        let registration_response = register_client(
+            &self.http_service(),
+            registration_endpoint,
+            client_metadata,
+            software_statement,
+        )
+        .await?;
 
         // The format of the credentials changes according to the client metadata that
         // was sent. Public clients only get a client ID.
@@ -996,6 +1063,24 @@ impl Oidc {
         Ok(())
     }
 
+    /// The scopes to request for logging in.
+    fn login_scopes(device_id: Option<String>) -> Result<[ScopeToken; 3], OidcError> {
+        // Generate the device ID if it is not provided.
+        let device_id = device_id.unwrap_or_else(|| {
+            rand::thread_rng()
+                .sample_iter(&rand::distributions::Alphanumeric)
+                .map(char::from)
+                .take(10)
+                .collect::<String>()
+        });
+
+        Ok([
+            ScopeToken::Openid,
+            ScopeToken::MatrixApi(MatrixApiScopeToken::Full),
+            ScopeToken::try_with_matrix_device(device_id).or(Err(OidcError::InvalidDeviceId))?,
+        ])
+    }
+
     /// Login via OpenID Connect with the Authorization Code flow.
     ///
     /// This should be called after [`Oidc::register_client()`] or
@@ -1072,24 +1157,7 @@ impl Oidc {
         redirect_uri: Url,
         device_id: Option<String>,
     ) -> Result<OidcAuthCodeUrlBuilder, OidcError> {
-        // Generate the device ID if it is not provided.
-        let device_id = if let Some(device_id) = device_id {
-            device_id
-        } else {
-            rand::thread_rng()
-                .sample_iter(&rand::distributions::Alphanumeric)
-                .map(char::from)
-                .take(10)
-                .collect::<String>()
-        };
-
-        let scope = [
-            ScopeToken::Openid,
-            ScopeToken::MatrixApi(MatrixApiScopeToken::Full),
-            ScopeToken::try_with_matrix_device(device_id).or(Err(OidcError::InvalidDeviceId))?,
-        ]
-        .into_iter()
-        .collect();
+        let scope = Self::login_scopes(device_id)?.into_iter().collect();
 
         Ok(OidcAuthCodeUrlBuilder::new(self.clone(), scope, redirect_uri))
     }
@@ -1187,17 +1255,22 @@ impl Oidc {
 
         let provider_metadata = self.provider_metadata().await?;
 
-        let session_tokens = self
-            .backend
-            .trade_authorization_code_for_tokens(
-                provider_metadata,
-                data.credentials(),
-                auth_code,
-                validation_data,
-            )
-            .await?;
+        let (response, _) = access_token_with_authorization_code(
+            &self.http_service(),
+            data.credentials(),
+            provider_metadata.token_endpoint(),
+            auth_code.code,
+            validation_data,
+            None,
+            Utc::now(),
+            &mut rng()?,
+        )
+        .await?;
 
-        self.set_session_tokens(session_tokens);
+        self.set_session_tokens(OidcSessionTokens {
+            access_token: response.access_token,
+            refresh_token: response.refresh_token,
+        });
 
         Ok(())
     }
@@ -1223,6 +1296,66 @@ impl Oidc {
         }
     }
 
+    /// Request codes from the authorization server for logging in with another
+    /// device.
+    #[cfg(all(feature = "e2e-encryption", not(target_arch = "wasm32")))]
+    async fn request_device_authorization(
+        &self,
+        device_id: Option<String>,
+    ) -> Result<oauth2::StandardDeviceAuthorizationResponse, qrcode::DeviceAuthorizationOauthError>
+    {
+        let scopes = Self::login_scopes(device_id)?
+            .into_iter()
+            .map(|scope| oauth2::Scope::new(scope.to_string()));
+
+        let client_id =
+            oauth2::ClientId::new(self.client_id().ok_or(OidcError::NotRegistered)?.0.clone());
+
+        let server_metadata = self.provider_metadata().await.map_err(OidcError::from)?;
+        let device_authorization_url = server_metadata
+            .device_authorization_endpoint
+            .clone()
+            .map(oauth2::DeviceAuthorizationUrl::from_url)
+            .ok_or(qrcode::DeviceAuthorizationOauthError::NoDeviceAuthorizationEndpoint)?;
+
+        let response = oauth2::basic::BasicClient::new(client_id)
+            .set_device_authorization_url(device_authorization_url)
+            .exchange_device_code()
+            .add_scopes(scopes)
+            .request_async(self.http_client())
+            .await?;
+
+        Ok(response)
+    }
+
+    /// Exchange the device code against an access token.
+    #[cfg(all(feature = "e2e-encryption", not(target_arch = "wasm32")))]
+    async fn exchange_device_code(
+        &self,
+        device_authorization_response: &oauth2::StandardDeviceAuthorizationResponse,
+    ) -> Result<(), qrcode::DeviceAuthorizationOauthError> {
+        use oauth2::TokenResponse;
+
+        let client_id =
+            oauth2::ClientId::new(self.client_id().ok_or(OidcError::NotRegistered)?.0.clone());
+
+        let server_metadata = self.provider_metadata().await.map_err(OidcError::from)?;
+        let token_uri = oauth2::TokenUrl::from_url(server_metadata.token_endpoint().clone());
+
+        let response = oauth2::basic::BasicClient::new(client_id)
+            .set_token_uri(token_uri)
+            .exchange_device_access_token(device_authorization_response)
+            .request_async(self.http_client(), tokio::time::sleep, None)
+            .await?;
+
+        self.set_session_tokens(OidcSessionTokens {
+            access_token: response.access_token().secret().to_owned(),
+            refresh_token: response.refresh_token().map(|t| t.secret().to_owned()),
+        });
+
+        Ok(())
+    }
+
     async fn refresh_access_token_inner(
         self,
         refresh_token: String,
@@ -1235,24 +1368,32 @@ impl Oidc {
             hash_str(&refresh_token)
         );
 
-        let new_tokens = self
-            .backend
-            .refresh_access_token(provider_metadata, credentials, refresh_token.clone())
-            .await?;
+        let (response, _) = refresh_access_token(
+            &self.http_service(),
+            credentials,
+            provider_metadata.token_endpoint(),
+            refresh_token.clone(),
+            None,
+            None,
+            None,
+            Utc::now(),
+            &mut rng()?,
+        )
+        .await?;
 
         trace!(
             "Token refresh: new refresh_token: {} / access_token: {:x}",
-            new_tokens
+            response
                 .refresh_token
                 .as_deref()
                 .map(|token| format!("{:x}", hash_str(token)))
                 .unwrap_or_else(|| "<none>".to_owned()),
-            hash_str(&new_tokens.access_token)
+            hash_str(&response.access_token)
         );
 
         let tokens = OidcSessionTokens {
-            access_token: new_tokens.access_token,
-            refresh_token: new_tokens.refresh_token.clone().or(Some(refresh_token)),
+            access_token: response.access_token,
+            refresh_token: response.refresh_token.or(Some(refresh_token)),
         };
 
         self.set_session_tokens(tokens.clone());
@@ -1417,14 +1558,16 @@ impl Oidc {
         let tokens = self.session_tokens().ok_or(OidcError::NotAuthenticated)?;
 
         // Revoke the access token, it should revoke both tokens.
-        self.backend
-            .revoke_token(
-                client_credentials,
-                revocation_endpoint,
-                tokens.access_token,
-                Some(OAuthTokenTypeHint::AccessToken),
-            )
-            .await?;
+        revoke_token(
+            &self.http_service(),
+            client_credentials,
+            revocation_endpoint,
+            tokens.access_token,
+            Some(OAuthTokenTypeHint::AccessToken),
+            Utc::now(),
+            &mut rng()?,
+        )
+        .await?;
 
         if let Some(manager) = self.ctx().cross_process_token_refresh_manager.get() {
             manager.on_logout().await?;
@@ -1669,4 +1812,19 @@ fn rng() -> Result<StdRng, OidcError> {
 
 fn hash_str(x: &str) -> impl fmt::LowerHex {
     sha2::Sha256::new().chain_update(x).finalize()
+}
+
+impl<'c> AsyncHttpClient<'c> for HttpClient {
+    type Error = HttpClientError<reqwest::Error>;
+
+    type Future =
+        Pin<Box<dyn Future<Output = Result<HttpResponse, Self::Error>> + Send + Sync + 'c>>;
+
+    fn call(&'c self, request: HttpRequest) -> Self::Future {
+        Box::pin(async move {
+            let response = self.inner.call(request).await?;
+
+            Ok(response)
+        })
+    }
 }
